Run feature tests only on purl crate
Using `--no-default-features` on the workspace does not work as expected
because the `purl_test` and `xtask` crates depend on certain features of
the `purl` crate. So with feature unification, the `purl` crate is built
with those features **enabled**.

To resolve this and simplify the tests, this patch switches the default
and minimal tests to only test the `purl` crate itself. The "Full" test
is left to test the entire workspace with code coverage.

This technically leaves open the possibility that `purl_test` or `xtask`
may not successfully build with default features... But since those are
test and development crates, I'm not actually concerned.